refactor: use aegir
